refactor(reports): split 816-line _data.py into _metrics + _html
Extract the metric-row builders into _reports/_metrics.py and the HTML
renderers (plus section-layout constants) into _reports/_html.py. The
Reports class stays in _data.py and imports both. Pure code movement —
public API unchanged, all gates green.
